+++ cmd/gardener-seed-admission-controller/app/gardener_seed_admission_controller.go
@@ -33,8 +33,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
 	"github.com/gardener/gardener/pkg/client/kubernetes"
-	"github.com/gardener/gardener/pkg/seedadmission"
 	"github.com/gardener/gardener/pkg/seedadmission/webhooks/admission/extensioncrds"
+	"github.com/gardener/gardener/pkg/seedadmission/webhooks/admission/podschedulername"
 )
 
 const (
@@ -151,10 +151,7 @@ func (o *Options) Run(ctx context.Context) error {
 	log.Info("setting up webhook server")
 	server := mgr.GetWebhookServer()
 	server.Register(extensioncrds.WebhookPath, &webhook.Admission{Handler: extensioncrds.New(logf.Log.WithName(extensioncrds.PluginName))})
-	server.Register(
-		seedadmission.GardenerShootControlPlaneSchedulerWebhookPath,
-		&webhook.Admission{Handler: admission.HandlerFunc(seedadmission.DefaultShootControlPlanePodsSchedulerName)},
-	)
+	server.Register(podschedulername.WebhookPath, &webhook.Admission{Handler: admission.HandlerFunc(podschedulername.DefaultShootControlPlanePodsSchedulerName)})
 
 	log.Info("starting manager")
 	if err := mgr.Start(ctx); err != nil {
